Separated cache logic from data access logic for getting contest categories https://github.com/SoftUni-Internal/exam-systems-issues/issues/130
