+++ Servers/UI/OJS.Servers.Ui/Controllers/Api/ContestCategoriesController.cs
@@ -6,21 +6,21 @@ using System.Threading.Tasks;
 
 public class ContestCategoriesController : ControllerBase
 {
-    private readonly IContestCategoriesBusinessService contestCategoriesBusiness;
+    private readonly IContestCategoriesCacheService contestCategoriesCache;
 
     public ContestCategoriesController(
-        IContestCategoriesBusinessService contestCategoriesBusiness)
-        => this.contestCategoriesBusiness = contestCategoriesBusiness;
+        IContestCategoriesCacheService contestCategoriesCache)
+        => this.contestCategoriesCache = contestCategoriesCache;
 
     public async Task<IActionResult> GetCategoriesTree()
-        => this.Ok(await this.contestCategoriesBusiness.GetTree());
+        => this.Ok(await this.contestCategoriesCache.GetAllContestCategoriesTree());
 
     public async Task<IActionResult> GetMainCategories()
-        => this.Ok(await this.contestCategoriesBusiness.GetAllMain());
+        => this.Ok(await this.contestCategoriesCache.GetMainContestCategories());
 
     public async Task<IActionResult> GetSubcategories(int id)
-        => this.Ok(await this.contestCategoriesBusiness.GetAllSubcategories(id));
+        => this.Ok(await this.contestCategoriesCache.GetContestSubCategoriesList(id));
 
     public async Task<IActionResult> GetParentCategories(int id)
-        => this.Ok(await this.contestCategoriesBusiness.GetAllParentCategories(id));
+        => this.Ok(await this.contestCategoriesCache.GetContestCategoryParentsList(id));
 }
