+++ Services/Administration/OJS.Services.Administration.Business/IContestCategoriesCacheService.cs
@@ -1,0 +1,11 @@
+namespace OJS.Services.Administration.Business;
+
+using SoftUni.Services.Infrastructure;
+using System.Threading.Tasks;
+
+public interface IContestCategoriesCacheService : IService
+{
+    void ClearMainContestCategoriesCache();
+
+    Task ClearContestCategory(int categoryId);
+}
