+++ Services/UI/OJS.Services.Ui.Data/IContestCategoriesDataService.cs
@@ -2,6 +2,7 @@
 {
     using OJS.Data.Models.Contests;
     using OJS.Services.Common.Data;
+    using System.Collections.Generic;
     using System.Linq;
     using System.Threading.Tasks;
 
@@ -14,5 +15,9 @@
         Task<string?> GetNameById(int id);
 
         Task<bool> HasContestsById(int id);
+
+        IQueryable<ContestCategory> GetAllVisibleOrdered();
+
+        Task<IEnumerable<TServiceModel>> GetAllVisibleMainOrdered<TServiceModel>();
     }
 }
